Clean up advertisement data
It made no sense to prepare the data a certain way in Objective-C to get re-prepared again on JavaScript to conform to the Noble API. Instead, the necessary data is prepared to look how we want in Objective-C and we let the React Native bridge take care of the data conversion. The only thing that JavaScript needs to do is Buffer conversion.